Update Readme. Complete prelim redistributePolygon
Readme.txt updated with instructions on pushing to git up in the face of
council's proxy.

Completed the redistributePolygon.py tool. This tool has produced
outputs which have been sent to Brian. There is still an issue with this
tool double counting large properties that span multiple zones. Possible
fix may be to use the centroid of the shape to redistribute the
population, but this would overcount some zone and undercount other.
If this issue exists with a significant number of properties, may need
to redistribute the properties somehow.