Workaround double counting issue
I have changed the way that the script redistributes the population of
the GMZs.To address the double counting issue, the total GMZ number of
properties is now recalculate from the intersecting_polygons layer.
This means that any properties that are double counted in adjacent
intersecting polygons, are recounted in the total number of
properties of the GMZ. Therefor, when population projections are
redistributed in terms of the ratio of properties, double counted
properties are taken into account. There will still exist issues of some
intersecting_polygons having a higher portion of population then they
should, but at least the summed total will be consistent. Though I have
not tested this and that will be the next step.

I have also worked to tidy up some of the code. This also needs a lot
more work.