Solve overestimation issue
Have added a function to return the max of the counted and the double
counted properties field. This fixes the issue of overestimating because
GMZ that weren't fully covered by intersecting polygons would have a
high portion of their population designated to a small intersecting
polygon.

Also cleaned up the code, changing variable names for clarity.