for_each_feature takes extra arguments
Any additional arguments passed to for_each_feature will be passed on to
the callback function. This was added for the Land_Use_Monitoring
project. It may have broken other tools because I think previously a
where_clause was also taken by this function.

I need to figure out how to handle this.