WIP: fix add_layer_count
When I changed add_layer_count to not overwrite the original layer, it
cause other problems. When the layer is copied, the OBJECTID changes and
can't be used for the join. Need to fix this, Not sure how to test it
though. Maybe I can use the files from distribute_growth to test as I
know they fail.

TODO: address the RuntimeError